Implement new connect account page design
Migrating it to TypeScript, and refactoring some affected code
