--- client/connect-account-page/info-modal/style.scss
@@ -1,21 +1,0 @@
-.woocommerce-payments__onboarding_more_info-modal {
-	&.components-modal__frame {
-		@media ( min-width: 600px ) {
-			max-width: 700px;
-		}
-	}
-
-	.wcpay-connect-warning-notice.components-notice {
-		margin: $gap 0 0;
-		padding: $gap-small $gap-small 0;
-	}
-
-	hr {
-		margin: $gap-large -32px 0;
-	}
-
-	.woocommerce-payments__onboarding_more_info-footer {
-		@include modal-footer-buttons;
-		margin-top: $gap-large;
-	}
-}
