+++ client/connect-account-page/style.scss
@@ -1,133 +1,156 @@
 .connect-account-page {
-	.wcpay-connect-account-page-masthead {
-		position: relative;
-		height: 120px;
-		border-bottom: 1px solid $studio-gray-5;
+	a {
+		text-decoration: none;
+	}
 
-		h1 {
-			position: absolute;
-			top: 24px;
-			left: 24px;
+	h2 {
+		@include wp-title-medium;
+		margin-bottom: $gap-smaller;
+	}
+
+	&__onboarding {
+		.components-button {
+			margin: $gap-smaller 0 0;
 		}
 
-		.wcpay-connect-account-page-recommended {
-			position: absolute;
-			top: 44px;
-			left: 246px;
-		}
+		.components-card__header {
+			img {
+				max-width: 100%;
+				height: 74px;
+				margin: 0 auto;
+			}
 
-		.wcpay-connect-account-page-hero {
-			@include breakpoint( '<660px' ) {
+			@media screen and ( min-width: $break-small ) {
 				display: none;
 			}
-			position: absolute;
-			top: -100px;
-			right: -36px;
-			clip: rect( 100px, 220px, 220px, auto );
 		}
 	}
 
-	.wcpay-connect-account-page-pill {
-		/* adapted from https://github.com/woocommerce/woocommerce-admin/blob/main/packages/components/src/pill/style.scss */
-		font-size: 12px;
-		border: 1px solid $studio-gray-70;
-		border-radius: 28px;
-		color: $studio-gray-70;
-		display: inline-block;
-		padding: 2px 8px;
-	}
-
-	.wcpay-connect-account-page-payment-methods {
-		img {
-			display: inline-block;
-			height: 35px;
-			width: 52px;
-			margin-right: 0.5em;
-			margin-bottom: 0.5em;
-			border: 1px solid #f3f3f3;
-		}
-	}
-
-	.wcpay-connect-account-page-terms-of-service {
-		display: block;
-		margin: 0.5em 0;
-	}
-
-	.connect-page-onboarding-steps {
-		@include breakpoint( '>660px' ) {
+	@media screen and ( min-width: $break-small ) {
+		div#{&}__content {
+			background: url( 'assets/images/illustrations/woopayments.svg?asset' )
+				no-repeat bottom right/293px 269px;
+			padding: $gap-largest;
+			padding-right: 340px;
+			min-height: 299px;
 			display: flex;
-			flex-direction: row;
+			flex-direction: column;
+			justify-content: flex-end;
+			align-items: flex-start;
 		}
+	}
 
-		.connect-page-onboarding-steps-item {
-			margin-top: 30px;
+	&__content {
+		p {
+			@include wc-others-caption;
+			color: $studio-gray-60;
+		}
+	}
 
-			@include breakpoint( '>660px' ) {
-				margin-top: 0;
-				margin-right: 80px;
+	div#{&}__payment-methods {
+		padding: $gap $gap-largest;
+		margin-bottom: $gap-smaller;
+
+		p {
+			text-align: center;
+			color: $studio-gray-40;
+			@media screen and ( min-width: $break-small ) {
+				text-align: initial;
 			}
 		}
-	}
 
-	.wcpay-connect-account-page-step-number {
-		background-color: $studio-woocommerce-purple-50;
-		color: $white;
-		font-size: 20px;
-		border-radius: 50%;
-		min-height: 32px;
-		min-width: 32px;
-		display: inline-block;
-		text-align: center;
-		line-height: 32px;
-	}
-
-	.connect-account {
-		h2 {
-			font-size: 20px;
-			font-weight: 300;
+		img {
+			height: 24px;
+			border: 1px solid $gray-300;
 		}
 
-		h3 {
-			font-size: 1em;
-		}
+		&__icons {
+			display: grid;
+			grid-gap: $gap-smaller;
+			grid-template-columns: repeat( auto-fill, minmax( 39px, 1fr ) );
 
-		.connect-account__card {
-			> .components-card__body {
-				padding: 0;
+			span {
+				@include wc-others-caption;
+				color: $gray-700;
+				display: none;
 
-				.content {
-					padding: 0 24px 24px 24px;
-
-					a {
-						text-decoration: none;
-					}
-
-					hr {
-						position: relative;
-						left: -24px;
-						width: calc( 100% + 48px );
-					}
+				@media screen and ( min-width: $break-small ) {
+					display: initial;
+					white-space: nowrap;
+					align-self: end;
 				}
 			}
 		}
+	}
 
-		.connect-account__action {
-			margin-top: 16px;
-			margin-bottom: 0;
+	&__details {
+		div.components-card__body {
+			padding: $gap-large;
 		}
 
-		.connect-account__steps {
-			.components-card__body {
-				padding: 24px;
-			}
-
-			h2 {
-				margin-top: 0;
-			}
+		.wcpay-component-tip-box {
+			margin: $gap-small 0 $gap-small;
 		}
 	}
 
-	.wcpay-component-tip-box {
-		margin-bottom: $gap-large;
+	&__steps {
+		@media screen and ( min-width: $break-small ) {
+			display: grid;
+			grid-template-columns: repeat( 3, 1fr );
+			column-gap: $gap-larger;
+			margin-left: -$gap-large;
+			margin-right: -$gap-large;
+		}
+	}
+
+	&__step {
+		padding: $gap-small 0;
+
+		@media screen and ( min-width: $break-small ) {
+			padding: $gap-large $gap-large 0 $gap-large;
+		}
+
+		span {
+			background-color: $studio-woocommerce-purple-50;
+			color: $white;
+			font-size: 20px;
+			border-radius: 50%;
+			min-height: 32px;
+			min-width: 32px;
+			display: inline-block;
+			text-align: center;
+			line-height: 32px;
+		}
+
+		h3 {
+			font-size: 12px;
+			line-height: 16px;
+			font-weight: 700;
+			color: $studio-gray-70;
+		}
+
+		p {
+			@include wc-body-small;
+			color: $studio-gray-50;
+		}
+	}
+
+	&__info-modal {
+		&.components-modal__frame {
+			max-width: 700px;
+		}
+
+		.components-notice {
+			margin: $gap 0 0;
+			padding: $gap-small $gap-small 0;
+		}
+
+		hr {
+			margin: $gap-large 0;
+		}
+
+		&__footer {
+			@include modal-footer-buttons;
+		}
 	}
 }
