+++ client/deposits/declarations.d.ts
@@ -32,6 +32,8 @@ declare module '@woocommerce/components' {
 		href: string;
 		children?: React.ReactNode;
 		type?: string;
+		target?: string;
+		rel?: string;
 		onClick?: ( ...props: any ) => any;
 	};
 	const Link: ( props: LinkParams ) => JSX.Element;
